[chore] Rename utils.py to scoring_primitives.py
